+++ core/Pos/Util/Util.hs
@@ -2,7 +2,6 @@
 {-# LANGUAGE ConstraintKinds #-}
 {-# LANGUAGE PolyKinds       #-}
 {-# LANGUAGE RankNTypes      #-}
-{-# LANGUAGE TypeFamilies    #-}
 
 module Pos.Util.Util
        (
@@ -55,242 +54,35 @@ module Pos.Util.Util
        -- * Aeson
        , parseJSONWithRead
 
-       -- * Instances
-       -- ** Lift Byte
-       -- ** Lift HashMap
-       -- ** FromJSON Byte, ToJSON Byte
-       -- ** Hashable Byte
-       -- ** MonadFail (Either s), assuming IsString s
-       -- ** HasLoggerName (MonadPseudoRandom drg)
-
-       -- ** Hashable
-       -- *** Millisecond, Microsecond
-
-       -- ** NFData
-       -- *** Millisecond, Microsecond
-
-       -- ** MonadRandom
-       -- *** monad transformers
-       -- *** Gen (from QuickCheck)
-
-       -- ** Buildable
-       -- *** "Data.Time.Units" types
-       -- *** @()@
        ) where
 
 import           Universum
 
 import           Control.Concurrent (myThreadId, threadDelay)
 import           Control.Lens (Getting, Iso', coerced, foldMapOf, ( # ))
-import           Control.Monad.Base (MonadBase)
 import qualified Control.Monad.Catch as MC
-import           Control.Monad.Morph (MFunctor (..))
 import           Control.Monad.Trans.Class (MonadTrans)
-import           Control.Monad.Trans.Identity (IdentityT (..))
-import           Control.Monad.Trans.Lift.Local (LiftLocal (..))
-import           Control.Monad.Trans.Resource (MonadResource (..), ResourceT, transResourceT)
-import qualified Crypto.Random as Rand
-import           Data.Aeson (FromJSON (..), ToJSON (..))
+import           Data.Aeson (FromJSON (..))
 import qualified Data.Aeson as A
 import qualified Data.Aeson.Types as A
 import           Data.Char (isAlphaNum)
-import           Data.Hashable (Hashable (hashWithSalt))
-import qualified Data.HashMap.Strict as HM
 import           Data.HashSet (fromMap)
 import           Data.List (last)
 import qualified Data.List.NonEmpty as NE
 import qualified Data.Map as M
 import qualified Data.Semigroup as Smg
-import           Data.Tagged (Tagged (Tagged))
-import           Data.Text.Buildable (build)
 import           Data.Time (getCurrentTime)
 import           Data.Time.Clock (NominalDiffTime)
-import           Data.Time.Units (Attosecond, Day, Femtosecond, Fortnight, Hour, Microsecond,
-                                  Millisecond, Minute, Nanosecond, Picosecond, Second, Week,
-                                  toMicroseconds)
-import           Data.Typeable (typeRep)
 import qualified Ether
 import           Ether.Internal (HasLens (..))
-import qualified Formatting as F
 import qualified Language.Haskell.TH as TH
-import qualified Language.Haskell.TH.Syntax as TH
-import           Mockable (ChannelT, Counter, Distribution, Gauge, MFunctor' (..), Mockable (..),
-                           Promise, SharedAtomicT, SharedExclusiveT, ThreadId)
 import qualified Prelude
-import           Serokell.Data.Memory.Units (Byte, fromBytes, toBytes)
 import qualified System.Console.ANSI as ANSI
 import           System.Directory (canonicalizePath, createDirectory, doesDirectoryExist,
                                    getTemporaryDirectory, listDirectory, removeDirectoryRecursive,
                                    removeFile)
 import           System.FilePath (normalise, pathSeparator, takeDirectory, (</>))
 import           System.IO (hClose, openTempFile)
-import           System.Wlog (CanLog, HasLoggerName (..), LoggerNameBox (..))
-import qualified Test.QuickCheck as QC
-import           Test.QuickCheck.Monadic (PropertyM (..))
-
-----------------------------------------------------------------------------
--- Instances
-----------------------------------------------------------------------------
-
-instance MonadReader r m => MonadReader r (PropertyM m) where
-    ask = lift ask
-    local f (MkPropertyM propertyM) =
-        MkPropertyM $ \hole -> local f <$> propertyM hole
-
-instance (TH.Lift k, TH.Lift v) => TH.Lift (HashMap k v) where
-    lift x = let l = HM.toList x in [|HM.fromList l|]
-
-instance Hashable Byte where
-    hashWithSalt i = hashWithSalt i . toInteger
-
-instance TH.Lift Byte where
-    lift x = let b = toBytes x in [|fromBytes b :: Byte|]
-
-instance FromJSON Byte where
-    parseJSON = fmap fromBytes . parseJSON
-
-instance ToJSON Byte where
-    toJSON = toJSON . toBytes
-
-instance IsString s => MonadFail (Either s) where
-    fail = Left . fromString
-
-instance Rand.DRG drg => HasLoggerName (Rand.MonadPseudoRandom drg) where
-    getLoggerName = pure "MonadPseudoRandom"
-    modifyLoggerName = flip const
-
-instance {-# OVERLAPPABLE #-}
-         (MonadTrans t, Functor (t m), Monad (t m), Rand.MonadRandom m)
-         => Rand.MonadRandom (t m) where
-    getRandomBytes = lift . Rand.getRandomBytes
-
-instance Rand.MonadRandom QC.Gen where
-    getRandomBytes n = do
-        [a,b,c,d,e] <- replicateM 5 (QC.choose (minBound, maxBound))
-        pure $ fst $ Rand.randomBytesGenerate n (Rand.drgNewTest (a,b,c,d,e))
-
-instance Hashable Millisecond where
-    hashWithSalt i a = hashWithSalt i (toInteger a)
-
-instance Hashable Microsecond where
-    hashWithSalt i a = hashWithSalt i (toInteger a)
-
-instance NFData Millisecond where
-    rnf ms = rnf (toInteger ms)
-
-instance NFData Microsecond where
-    rnf ms = rnf (toInteger ms)
-
-
-----------------------------------------------------------------------------
--- Orphan Buildable instances
-----------------------------------------------------------------------------
-
-instance Buildable Attosecond  where build = build @String . show
-instance Buildable Femtosecond where build = build @String . show
-instance Buildable Picosecond  where build = build @String . show
-instance Buildable Nanosecond  where build = build @String . show
-instance Buildable Millisecond where build = build @String . show
-instance Buildable Second      where build = build @String . show
-instance Buildable Minute      where build = build @String . show
-instance Buildable Hour        where build = build @String . show
-instance Buildable Day         where build = build @String . show
-instance Buildable Week        where build = build @String . show
-instance Buildable Fortnight   where build = build @String . show
-
--- | Special case. We don't want to print greek letter mu in console because
--- it breaks things sometimes.
-instance Buildable Microsecond where
-    build = build . (++ "mcs") . show . toMicroseconds
-
--- | This orphan instance is sometimes useful and why not have it?
-instance Buildable () where
-    build _ = "()"
-
-----------------------------------------------------------------------------
--- MonadResource/ResourceT
-----------------------------------------------------------------------------
-
-instance LiftLocal ResourceT where
-    liftLocal _ l f = hoist (l f)
-
-instance {-# OVERLAPPABLE #-}
-    (MonadResource m, MonadTrans t, Applicative (t m),
-     MonadBase IO (t m), MonadIO (t m), MonadThrow (t m)) =>
-        MonadResource (t m)
-  where
-    liftResourceT = lift . liftResourceT
-
--- TODO Move it to log-warper
-instance CanLog m => CanLog (ResourceT m)
-instance (Monad m, HasLoggerName m) => HasLoggerName (ResourceT m) where
-    getLoggerName = lift getLoggerName
-    modifyLoggerName = transResourceT . modifyLoggerName
-
-----------------------------------------------------------------------------
--- Instances required by 'ether'
-----------------------------------------------------------------------------
-
-instance
-    (Monad m, MonadTrans t, Monad (t m), CanLog m) =>
-        CanLog (Ether.TaggedTrans tag t m)
-
-instance (Monad m, CanLog m) => CanLog (IdentityT m)
-
-instance
-    (LiftLocal t, Monad m, HasLoggerName m) =>
-        HasLoggerName (Ether.TaggedTrans tag t m)
-  where
-    getLoggerName = lift getLoggerName
-    modifyLoggerName = liftLocal getLoggerName modifyLoggerName
-
-instance
-    (Monad m, HasLoggerName m) => HasLoggerName (IdentityT m)
-  where
-    getLoggerName = lift getLoggerName
-    modifyLoggerName = liftLocal getLoggerName modifyLoggerName
-
-deriving instance LiftLocal LoggerNameBox
-
-instance {-# OVERLAPPABLE #-}
-    (Monad m, MFunctor t) => MFunctor' t m n
-  where
-    hoist' = hoist
-
-instance
-    (Mockable d m, MFunctor' d (IdentityT m) m) =>
-        Mockable d (IdentityT m)
-  where
-    liftMockable dmt = IdentityT $ liftMockable $ hoist' runIdentityT dmt
-
-unTaggedTrans :: Ether.TaggedTrans tag t m a -> t m a
-unTaggedTrans (Ether.TaggedTrans tma) = tma
-
-instance
-      (Mockable d (t m), Monad (t m),
-       MFunctor' d (Ether.TaggedTrans tag t m) (t m)) =>
-          Mockable d (Ether.TaggedTrans tag t m)
-  where
-    liftMockable dmt =
-      Ether.TaggedTrans $ liftMockable $ hoist' unTaggedTrans dmt
-
-type instance ThreadId (IdentityT m) = ThreadId m
-type instance Promise (IdentityT m) = Promise m
-type instance SharedAtomicT (IdentityT m) = SharedAtomicT m
-type instance Counter (IdentityT m) = Counter m
-type instance Distribution (IdentityT m) = Distribution m
-type instance SharedExclusiveT (IdentityT m) = SharedExclusiveT m
-type instance Gauge (IdentityT m) = Gauge m
-type instance ChannelT (IdentityT m) = ChannelT m
-
-type instance ThreadId (Ether.TaggedTrans tag t m) = ThreadId m
-type instance Promise (Ether.TaggedTrans tag t m) = Promise m
-type instance SharedAtomicT (Ether.TaggedTrans tag t m) = SharedAtomicT m
-type instance Counter (Ether.TaggedTrans tag t m) = Counter m
-type instance Distribution (Ether.TaggedTrans tag t m) = Distribution m
-type instance SharedExclusiveT (Ether.TaggedTrans tag t m) = SharedExclusiveT m
-type instance Gauge (Ether.TaggedTrans tag t m) = Gauge m
-type instance ChannelT (Ether.TaggedTrans tag t m) = ChannelT m
 
 ----------------------------------------------------------------------------
 -- Not instances
@@ -355,12 +147,6 @@ instance {-# OVERLAPPING #-} PowerLift m m where
 
 instance (MonadTrans t, PowerLift m n, Monad n) => PowerLift m (t n) where
   powerLift = lift . powerLift @m @n
-
-instance (Typeable s, Buildable a) => Buildable (Tagged s a) where
-    build tt@(Tagged v) = F.bprint ("Tagged " F.% F.shown F.% " " F.% F.build) ts v
-      where
-        ts = typeRep proxy
-        proxy = (const Proxy :: Tagged s a -> Proxy s) tt
 
 -- | This function performs checks at compile-time for different actions.
 -- May slowdown implementation. To disable such checks (especially in benchmarks)
