+++ explorer/cardano-sl-explorer.cabal
@@ -88,6 +88,7 @@ library
 
                       -- cardano sl
                       , cardano-sl
+                      , cardano-sl-util
                       , cardano-sl-core
                       , cardano-sl-infra
                       , cardano-sl-txp
