+++ util/cardano-sl-util.cabal
@@ -29,19 +29,23 @@ library
                        Pos.Util.Justify
                        Pos.Util.Lens
                        Pos.Util.LRU
+                       Pos.Util.Orphans
                        Pos.Util.Queue
                        Pos.Util.Some
 
   other-modules:
 
   build-depends:       QuickCheck
+                     , aeson
                      , autoexporter
                      , base
                      , bytestring
                      , containers
                      , concurrent-extra
+                     , cryptonite
                      , data-default
                      , deepseq
+                     , ether
                      , exceptions
                      , formatting
                      , hashable
@@ -50,13 +54,18 @@ library
                      , lrucache
                      , mmorph
                      , mtl
+                     , node-sketch
+                     , parsec
                      , random
+                     , resourcet
                      , semigroups
                      , serokell-util
                      , stm
+                     , tagged
                      , template-haskell
                      , text
                      , text-format
+                     , time-units
                      , transformers
                      , transformers-base
                      , transformers-lift
